+++ config/routes.rb
@@ -31,6 +31,7 @@ Rails.application.routes.draw do
       post "projects/:authentication_token/subjects/:id/sheets", to: "subjects#create_sheet", as: :create_sheet
 
       get "projects/:authentication_token/subjects/:id/surveys/:event/:design", to: "surveys#show_survey", as: :show_survey
+      get "projects/:authentication_token/subjects/:id/surveys/:event/:design/resume", to: "surveys#resume_survey", as: :resume_survey
       get "projects/:authentication_token/subjects/:id/surveys/:event/:design/:page", to: "surveys#show_survey_page", as: :show_survey_page
       patch "projects/:authentication_token/subjects/:id/surveys/:event/:design/:page", to: "surveys#update_survey_response", as: :update_survey_response
     end